+++ cpp_yandex/courses/course1/lesson2/main7.cpp
@@ -7,7 +7,7 @@ vector<int>
 Reversed(const vector<int>& v)
 {
 	vector<int> res;
-	for (int i = v.size() - 1; i >= 0; i--)
+	for (long long i = v.size() - 1; i >= 0; i--)
 		res.push_back(v[i]);
 	return res;
 }
